+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/docsig/compare/v0.38.0...HEAD)
 ------------------------------------------------------------------------
+### Added
+- add function name to summary report header
+
 ### Changed
 - sort errors for report output